+++ r/draw_sbgnml.R
@@ -6,6 +6,8 @@
 # LOAD DATA ----
 
 suppressPackageStartupMessages(library(jsonlite))
+suppressPackageStartupMessages(library(showtext))
+suppressPackageStartupMessages(library(sysfonts))
 suppressPackageStartupMessages(library(xml2))
 
 arguments <- commandArgs(trailingOnly = FALSE)
@@ -15,6 +17,7 @@ script_dir <- if (length(file_argument) == 0) {
 } else {
   dirname(normalizePath(sub("^--file=", "", file_argument[1]), mustWork = TRUE))
 }
+options(renderSbgnR.source_root = script_dir)
 source(file.path(script_dir, "R", "draw_sbgnml.R"))
 
 # ANALYSIS ----